Added Custom buttons and custom combo boxes.
The combo  boxes had issues of white borders. This should get rid of the white borders.

The keyboard buttons had issues of not keeping selection between mode switches, this should fix that.

There is a issue where a random button would be considered selected at the start of the program,  I should figure something out for that.